moved notif folder out of app into src
