+++ src/notifications/helper-functions/geoHashHelpers.ts
