+++ src/notifications/helper-functions/unsocialHoursCalculationHelpers.tsx
@@ -1,6 +1,6 @@
 import { DateTime } from "luxon";
 import { NotifUnsocialHours } from "../types/types";
-import { UNSOCIAL_HOUR_END, UNSOCIAL_HOUR_START } from "../../../app/constants";
+import { UNSOCIAL_HOUR_END, UNSOCIAL_HOUR_START } from "../../app/constants";
 
 const isUnsocialHour = (hour: number): boolean => {
     return hour < UNSOCIAL_HOUR_END || hour >= UNSOCIAL_HOUR_START;
